made some more changes to things
