+++ mymixtape-api/models/spotify.go
@@ -1,1 +1,299 @@
 package models
+
+// Spotify API Requests
+
+// Playlists
+type SpotifyAddItemsToPlaylistRequest struct {
+	URIs		[]string	`json:"uris"`
+	Position	int			`json:"position"`
+}
+
+type SpotifyCreatePlaylistRequest struct {
+	Name			string	`json:"name"`
+	Public			bool	`json:"public"`
+	Collaborative	bool	`json:"collaborative"`
+	Description		string	`json:"description"`
+}
+
+// Spotify API Responses
+// Authorization
+type SpotifyAccessTokenResponse struct {
+	AccessToken		string	`json:"access_token"`
+	TokenType		string	`json:"token_type"`
+	ExpiresIn		int		`json:"expires_in"`
+}
+
+type SpotifyAuthorizationUrlResponse struct {
+	Url	string	`json:"url"`
+}
+
+// User Profile
+type SpotifyCurrentUsersProfileResponse struct {
+	Country         string `json:"country"`
+	DisplayName     string `json:"display_name"`
+	Email           string `json:"email"`
+	ExplicitContent struct {
+		FilterEnabled bool `json:"filter_enabled"`
+		FilterLocked  bool `json:"filter_locked"`
+	} `json:"explicit_content"`
+	ExternalUrls struct {
+		Spotify string `json:"spotify"`
+	} `json:"external_urls"`
+	Followers struct {
+		Href  string `json:"href"`
+		Total int    `json:"total"`
+	} `json:"followers"`
+	Href   string `json:"href"`
+	ID     string `json:"id"`
+	Images []struct {
+		URL    string `json:"url"`
+		Height int    `json:"height"`
+		Width  int    `json:"width"`
+	} `json:"images"`
+	Product string `json:"product"`
+	Type    string `json:"type"`
+	URI     string `json:"uri"`
+}
+
+// User Playlists
+type SpotifyCurrentUsersPlaylistsResponse struct {
+	Href     string `json:"href"`
+	Limit    int    `json:"limit"`
+	Next     string `json:"next"`
+	Offset   int    `json:"offset"`
+	Previous string `json:"previous"`
+	Total    int    `json:"total"`
+	Items    []struct {
+		Collaborative bool   `json:"collaborative"`
+		Description   string `json:"description"`
+		ExternalUrls  struct {
+			Spotify string `json:"spotify"`
+		} `json:"external_urls"`
+		Href   string `json:"href"`
+		ID     string `json:"id"`
+		Images []struct {
+			URL    string `json:"url"`
+			Height int    `json:"height"`
+			Width  int    `json:"width"`
+		} `json:"images"`
+		Name  string `json:"name"`
+		Owner struct {
+			ExternalUrls struct {
+				Spotify string `json:"spotify"`
+			} `json:"external_urls"`
+			Followers struct {
+				Href  string `json:"href"`
+				Total int    `json:"total"`
+			} `json:"followers"`
+			Href        string `json:"href"`
+			ID          string `json:"id"`
+			Type        string `json:"type"`
+			URI         string `json:"uri"`
+			DisplayName string `json:"display_name"`
+		} `json:"owner"`
+		Public     bool   `json:"public"`
+		SnapshotID string `json:"snapshot_id"`
+		Tracks     struct {
+			Href  string `json:"href"`
+			Total int    `json:"total"`
+		} `json:"tracks"`
+		Type string `json:"type"`
+		URI  string `json:"uri"`
+	} `json:"items"`
+}
+
+
+
+type SpotifyAddItemsToPlaylistResponse struct {
+	SnapshotID	string	`json:"snapshot_id"`
+}
+
+type SpotifyPlaylistItemsResponse struct {
+	Href     string 	`json:"href"`
+	Limit    int    	`json:"limit"`
+	Next     string    	`json:"next"`
+	Offset   int    	`json:"offset"`
+	Previous string    	`json:"previous"`
+	Total    int    	`json:"total"`
+	Items    []struct {
+		Track struct {
+			ID   string `json:"id"`
+			Name string `json:"name"`
+		} `json:"track"`
+	} `json:"items"`
+}
+
+type SpotifyCreatePlaylistResponse struct {
+	Collaborative bool   `json:"collaborative"`
+	Description   string `json:"description"`
+	ExternalUrls  struct {
+		Spotify string `json:"spotify"`
+	} `json:"external_urls"`
+	Followers struct {
+		Href  string `json:"href"`
+		Total int    `json:"total"`
+	} `json:"followers"`
+	Href   string `json:"href"`
+	ID     string `json:"id"`
+	Images []struct {
+		URL    string `json:"url"`
+		Height int    `json:"height"`
+		Width  int    `json:"width"`
+	} `json:"images"`
+	Name  string `json:"name"`
+	Owner struct {
+		ExternalUrls struct {
+			Spotify string `json:"spotify"`
+		} `json:"external_urls"`
+		Followers struct {
+			Href  string `json:"href"`
+			Total int    `json:"total"`
+		} `json:"followers"`
+		Href        string `json:"href"`
+		ID          string `json:"id"`
+		Type        string `json:"type"`
+		URI         string `json:"uri"`
+		DisplayName string `json:"display_name"`
+	} `json:"owner"`
+	Public     bool   `json:"public"`
+	SnapshotID string `json:"snapshot_id"`
+	Tracks     struct {
+		Href     string `json:"href"`
+		Limit    int    `json:"limit"`
+		Next     string `json:"next"`
+		Offset   int    `json:"offset"`
+		Previous string `json:"previous"`
+		Total    int    `json:"total"`
+		Items    []struct {
+			AddedAt string `json:"added_at"`
+			AddedBy struct {
+				ExternalUrls struct {
+					Spotify string `json:"spotify"`
+				} `json:"external_urls"`
+				Followers struct {
+					Href  string `json:"href"`
+					Total int    `json:"total"`
+				} `json:"followers"`
+				Href string `json:"href"`
+				ID   string `json:"id"`
+				Type string `json:"type"`
+				URI  string `json:"uri"`
+			} `json:"added_by"`
+			IsLocal bool `json:"is_local"`
+			Track   struct {
+				Album struct {
+					AlbumType        string   `json:"album_type"`
+					TotalTracks      int      `json:"total_tracks"`
+					AvailableMarkets []string `json:"available_markets"`
+					ExternalUrls     struct {
+						Spotify string `json:"spotify"`
+					} `json:"external_urls"`
+					Href   string `json:"href"`
+					ID     string `json:"id"`
+					Images []struct {
+						URL    string `json:"url"`
+						Height int    `json:"height"`
+						Width  int    `json:"width"`
+					} `json:"images"`
+					Name                 string `json:"name"`
+					ReleaseDate          string `json:"release_date"`
+					ReleaseDatePrecision string `json:"release_date_precision"`
+					Restrictions         struct {
+						Reason string `json:"reason"`
+					} `json:"restrictions"`
+					Type       string `json:"type"`
+					URI        string `json:"uri"`
+					Copyrights []struct {
+						Text string `json:"text"`
+						Type string `json:"type"`
+					} `json:"copyrights"`
+					ExternalIds struct {
+						Isrc string `json:"isrc"`
+						Ean  string `json:"ean"`
+						Upc  string `json:"upc"`
+					} `json:"external_ids"`
+					Genres     []string `json:"genres"`
+					Label      string   `json:"label"`
+					Popularity int      `json:"popularity"`
+					AlbumGroup string   `json:"album_group"`
+					Artists    []struct {
+						ExternalUrls struct {
+							Spotify string `json:"spotify"`
+						} `json:"external_urls"`
+						Href string `json:"href"`
+						ID   string `json:"id"`
+						Name string `json:"name"`
+						Type string `json:"type"`
+						URI  string `json:"uri"`
+					} `json:"artists"`
+				} `json:"album"`
+				Artists []struct {
+					ExternalUrls struct {
+						Spotify string `json:"spotify"`
+					} `json:"external_urls"`
+					Followers struct {
+						Href  string `json:"href"`
+						Total int    `json:"total"`
+					} `json:"followers"`
+					Genres []string `json:"genres"`
+					Href   string   `json:"href"`
+					ID     string   `json:"id"`
+					Images []struct {
+						URL    string `json:"url"`
+						Height int    `json:"height"`
+						Width  int    `json:"width"`
+					} `json:"images"`
+					Name       string `json:"name"`
+					Popularity int    `json:"popularity"`
+					Type       string `json:"type"`
+					URI        string `json:"uri"`
+				} `json:"artists"`
+				AvailableMarkets []string `json:"available_markets"`
+				DiscNumber       int      `json:"disc_number"`
+				DurationMs       int      `json:"duration_ms"`
+				Explicit         bool     `json:"explicit"`
+				ExternalIds      struct {
+					Isrc string `json:"isrc"`
+					Ean  string `json:"ean"`
+					Upc  string `json:"upc"`
+				} `json:"external_ids"`
+				ExternalUrls struct {
+					Spotify string `json:"spotify"`
+				} `json:"external_urls"`
+				Href       string `json:"href"`
+				ID         string `json:"id"`
+				IsPlayable bool   `json:"is_playable"`
+				LinkedFrom struct {
+				} `json:"linked_from"`
+				Restrictions struct {
+					Reason string `json:"reason"`
+				} `json:"restrictions"`
+				Name        string `json:"name"`
+				Popularity  int    `json:"popularity"`
+				PreviewURL  string `json:"preview_url"`
+				TrackNumber int    `json:"track_number"`
+				Type        string `json:"type"`
+				URI         string `json:"uri"`
+				IsLocal     bool   `json:"is_local"`
+			} `json:"track"`
+		} `json:"items"`
+	} `json:"tracks"`
+	Type string `json:"type"`
+	URI  string `json:"uri"`
+}
+
+
+// Error Responses
+type SpotifyAuthenticationErrorResponse struct {
+	Error	string	`json:"error"`
+	Description string	`json:"error_description"`
+}
+
+type SpotifyErrorResponse struct {
+	Error	SpotifyErrorObjectResponse	`json:"error"`
+}
+
+type SpotifyErrorObjectResponse struct {
+	Status	int		`json:"status"`
+	Message	string	`json:"message"`
+}
