+++ mymixtape-api/server.go
@@ -7,5 +7,4 @@ func main() {
 	engine := router.InitRoutes()
 
 	panic(engine.Run())
-
 }
